fix: internationalize manifest.json
fixes #2020
