+++ .dockerignore
@@ -14,6 +14,7 @@ tests
 /static/icons.svg
 /static/inline-script.js.map
 /static/emoji-*.json
+/static/manifest.json
 /src/inline-script/checksum.js
 yarn-error.log
 /vercel.json
