+++ .vercelignore
@@ -8,5 +8,6 @@
 /static/icons.svg
 /static/inline-script.js.map
 /static/emoji-*.json
+/static/manifest.json
 /src/inline-script/checksum.js
 yarn-error.log
