--- static/manifest.webapp
@@ -1,37 +1,0 @@
-{
-  "version": "1.0.0",
-  "name": "Pinafore",
-  "description": "Mastodon client",
-  "type": "web",
-  "launch_path": "/",
-  "icons": {
-    "48": "/icon-48.png",
-    "72": "/icon-72.png",
-    "96": "/icon-96.png"
-  },
-  "developer": {
-    "name": "Nolan Lawson",
-    "url": "https://nolanlawson.com"
-  },
-  "locales": {
-    "en-US": {
-      "name": "Pinafore",
-      "subtitle": "Mastodon client",
-      "description": "Mastodon client, focused on speed and simplicity."
-    }
-  },
-  "default_locale": "en-US",
-  "permissions": {
-    "serviceworker": {
-      "description": "Needed for offline storage and notifications."
-    },
-    "desktop-notification": {
-      "description": "Needed for notifications."
-    }
-  },
-  "messages": [
-    {
-      "serviceworker-notification": "/"
-    }
-  ]
-}
